--- a/workspaces/rbac/.changeset/spicy-chairs-allow.md
+++ b/workspaces/rbac/.changeset/spicy-chairs-allow.md
@@ -1,0 +1,5 @@
+---
+'@backstage-community/plugin-rbac-backend': minor
+---
+
+Improve rbac performance using multi-value permission filters in the enforcer-delegate.ts.