## Hey, I just made a Pull Request!

Improve rbac performance using multi-value permission filters in the enforcer-delegate.ts

## How to test it

You can use my python script to messuare permission evaluation latency for user assigned to a lot of roles: https://github.com/PatAKnight/catalog-rbac-tests/pull/10

Result from the my laptop:

main branch [commit](https://github.com/backstage/community-plugins/commit/08f71c5eb755c2b4b67ef715744f65f550f30ed0):

```
Max latency value is 0.107 seconds
Min latency value is 0.055 seconds
Average request duration is time: 0.064 seconds
```

And my pr rebase to the same [commit](https://github.com/backstage/community-plugins/commit/08f71c5eb755c2b4b67ef715744f65f550f30ed0):

```
Max latency value is 0.082 seconds
Min latency value is 0.037 seconds
Average request duration is time: 0.042 seconds
```

I used user with generated 50 roles and script made default 1000 requests.

So (((0.064 - 0.042) / 0.064) * 100) = 34,375 %

#### :heavy_check_mark: Checklist

<!--- Please include the following in your Pull Request when applicable: -->

- [ ] A changeset describing the change and affected packages. ([more info](https://github.com/backstage/community-plugins/blob/master/CONTRIBUTING.md#creating-changesets))
- [ ] Added or updated documentation
- [ ] Tests for new functionality and regression tests for bug fixes
- [ ] Screenshots attached (for UI changes)
- [ ] All your commits have a `Signed-off-by` line in the message. ([more info](https://github.com/backstage/community-plugins/blob/master/CONTRIBUTING.md#developer-certificate-of-origin))